Reorganisation of parameterisation into separate subdirectory
