+++ ocelot/transformations/__init__.py
@@ -4,15 +4,5 @@ from .cleanup import (
     drop_zero_pv_row_datasets,
     ensure_all_datasets_have_production_volume,
 )
-from .known_ecoinvent_issues import fix_known_ecoinvent_issues
-from .parameterization import (
-    parameterization_validity_checks,
-    parameter_names_are_unique,
-)
+from .parameterization import fix_ecoinvent_parameters, recalculate
 
-
-def dummy_transformation(data):
-    """This is a dummy transformation that doesn't do anything.
-
-    Used primarily for testing."""
-    return data
