+++ tests/transformations/utils_tests.py
@@ -277,3 +277,44 @@ def test_choose_reference_product_exchange_zero_production(no_normalization):
     }]}
     with pytest.raises(ZeroProduction):
         choose_reference_product_exchange(given, given['exchanges'][0])
+
+@pytest.fixture(scope="function")
+def parameterized_ds():
+    return {
+        'exchanges': [{
+            'amount': 3.1415926535,
+            'variable': 'pie',
+            'production volume': {  # Nonsensical but should work
+                'variable': 'number_blueberries',
+                'amount': 42
+            },
+            'properties': [{
+                'variable': 'blueberry_volume',
+                'amount': 17
+            }]
+        }, {
+            'variable': 'circle',
+            'formula': 'pie * radius ** 2',
+            'properties': [{
+                'variable': 'radius',
+                'formula': 'blueberry_size * number_blueberries'
+            }]
+        }],
+        'parameters': [{
+            'variable': 'blueberry_size',
+            'formula': 'blueberry_density * blueberry_volume'
+        }, {
+            'variable': 'blueberry_density',
+            'amount': 1
+        }]
+    }
+
+def test_iterate_all_parameters(parameterized_ds):
+    generator = iterate_all_parameters(parameterized_ds)
+    assert next(generator) == parameterized_ds['exchanges'][0]
+    assert next(generator) == parameterized_ds['exchanges'][0]['production volume']
+    assert next(generator) == parameterized_ds['exchanges'][0]['properties'][0]
+    assert next(generator) == parameterized_ds['exchanges'][1]
+    assert next(generator) == parameterized_ds['exchanges'][1]['properties'][0]
+    assert next(generator) == parameterized_ds['parameters'][0]
+    assert next(generator) == parameterized_ds['parameters'][1]
